Add full inversions test

+++ inversions.go
@@ -1,7 +1,7 @@
 package main
 
-func mergeAndCount(leftArray, rightArray []int) (int, []int) {
-	var array []int
+func mergeAndCount(leftArray, rightArray []int64) (int, []int64) {
+	var array []int64
 	var i, j, inversions int
 	for i < len(leftArray) && j < len(rightArray) {
 		if leftArray[i] <= rightArray[j] {
@@ -19,14 +19,14 @@ func mergeAndCount(leftArray, rightArray []int) (int, []int) {
 	return inversions, append(array, leftArray[i:]...)
 }
 
-func sortAndCount(array []int) (int, []int) {
+func sortAndCount(array []int64) (int, []int64) {
 	if array[0] >= array[1] {
-		return 1, []int{array[1], array[0]}
+		return 1, []int64{array[1], array[0]}
 	}
 	return 0, array
 }
 
-func countInversions(inversions int, array []int) (int, []int) {
+func countInversions(inversions int, array []int64) (int, []int64) {
 	if len(array) > 2 {
 		half := len(array) / 2
 		leftInversions, leftArray := countInversions(inversions, array[:half])
